Add jsonschema validation for generated json
Move the stage2.schema.json from the root of the project into the src
directory, include it explicitly in MANIFEST.in to have it included
in the resulting nvme-spex package.

Add jsonschema as dependency.

Add cmdline flag `--validate-json`. When running spex with the new flag
it will validate that the resulting json aligns with the schema.