refactor(parser): make the parser testable
It was previously not possible to feed the json parser just
the raw html string but it was enforced to call it with a path
to a file on the fs.

This makes it exceedingly difficult to create any tests that
execute the parser.

This new way of splitting and wrapping the file read and parse
makes this easier.

Signed-off-by: Morten B. Rasmussen <m.rasmussen@samsung.com>