can_apply overhaul
Change the way extractors can determine if they are applicable to a
table (given only the table column headers).
Previously the strategy was hard-coded, for each type of information the
extractor was interested in (a semantic component), such as labels for
fields, values in case of enums or bit-/byte- ranges for struct tables,
the figure type would have a prioritized list of column headers to look
for.

This generally works well, but falls a bit short in case of value tables
where column names vary a lot.
One case, which we will like to solve, is identifying value tables where
either the 'value' or 'definition' column is found, and where the table
only has two columns (and neither is bits/bytes).
In this case, we can safely assume the table is a value table and the
other column is effectively the component we didn't identify.

Further changes are made to make overriding table extractors easier.
Each table extractor now implements `_can_apply`, which is called by
`can_apply`.
`_can_apply` is supposed to return a `Mapping` type, which maps
from a semantic component (i.e. 'label_column') to a concrete table
header for the table in question (i.e. 'name' in some cases,
'description' in others, and so on).

Components for which the extractor failed to find a mapping are set to
None.
This approach makes it easy to inherit from an extractor, for example to
override a particular figure, and build upon the mappings the base class
already found, only altering or adding the missing/misidentified
mappings.

The point of `can_apply` is that it correctly enumerates all fields of a
mapping type and returns None if *any* of the mapping entries are set to
None (not identified).

NOTE: each step was verified to cause no changes in linting errors
across all documents.

Signed-off-by: Jesper Wendel Devantier <j.devantier@samsung.com>