fix(ci): aggregate release workflow into build
The current status is that the release job is not run only when creating
a release. After looking into the json data in the github.event data it
does not look like the event data contains any reference to release or
tags when it is triggered by the workflow_run event.

This change will aggregate the release job and build job into the build
workflow.
To ensure that the release job is only triggered on the main branch when
and only when the build job has concluded successfully and a new tag has
been published. It further depends on the tests job to be successfully
also. See the `needs` attribute.