fix(type): add fields and remove data field on `Entity`
A search throughout the code base seems to show that the `data` entry of
the Entity is never set or nor accessed. However the entry is missing
the `fields` entry.

The new entry is a list, either type StructField or ValueField. To be
able to reference these in the `Entity` it is needed to move the bit of
code. Such that it is defined before use.

Signed-off-by: Morten B. Rasmussen <m.rasmussen@samsung.com>