feat(ui): new reporting view for linting errors
Builds upon the existing report view but now includes the html part of
the figures that have linting violations.

Further there was an issue with showing a limited amount of linting
errors. The functionality that should show a list of linting errors, was
iterating through the list of parsed figures and appending its linting
errors. This means that tables that are not being parsed by error or
otherwise would not be shown.  Now It iterates through the errors and
show that.

Signed-off-by: Morten B. Rasmussen <m.rasmussen@samsung.com>